moved util class to core package (as it is not part of "official" API)
