+++ bundles/automation/org.eclipse.smarthome.automation.api/META-INF/MANIFEST.MF
@@ -21,5 +21,4 @@ Export-Package:
  org.eclipse.smarthome.automation.provider,
  org.eclipse.smarthome.automation.provider.util,
  org.eclipse.smarthome.automation.template,
- org.eclipse.smarthome.automation.type,
- org.eclipse.smarthome.automation.util
+ org.eclipse.smarthome.automation.type
